requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
openpyxl
# python-docx
BeautifulSoup4
altair
requests
pandas
# selenium
# %LOCALAPPDATA%\Temp\gen_py